Adding previous and next buttons to result page.

+++ server/src/api/search/models/search.rs
@@ -9,14 +9,14 @@ use serde::{
 pub struct SearchQuery {
     pub query : String,
     pub preferred_instance : String,
-    pub page : Option<usize>
+    pub page : Option<i32>
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone, Default)]
 pub struct SearchResult {
     pub original_query_terms : HashSet<String>,
-    pub total_results : usize,
-    pub total_pages : usize,
+    pub total_results : i32,
+    pub total_pages : i32,
     pub time_taken : Duration,
     pub posts : Vec<SearchPost>
 }
